cleanup – merged std io
